Python-Works/DjangoApps/api/requirements.txt
altgraph==0.17.3
asgiref==3.5.2
asttokens==2.1.0
async-generator==1.10
attrs==22.1.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chatGPT==0.3.6
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
constantly==15.1.0
contourpy==1.0.6
cryptography==38.0.4
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.6
distro==1.8.0
Django==4.1.3
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
djangorestframework==3.14.0
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.4
executing==1.2.0
filelock==3.8.2
Flask==2.2.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
fonttools==4.38.0
future==0.18.2
googletrans==3.0.0
greenlet==2.0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
ipykernel==6.17.0
ipython==8.6.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==7.4.4
jupyter_core==4.11.2
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mysqlclient==2.1.1
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.23.4
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
pandas==1.5.2
parsel==1.7.0
parso==0.8.3
passlib==1.7.4
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
playwright==1.29.1
plotly==5.11.0
prompt-toolkit==3.0.32
Protego==0.2.1
psutil==5.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyee==9.0.4
Pygments==2.13.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyPDF2==3.0.1
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
pywin32==305
pywin32-ctypes==0.2.0
pyzmq==24.0.1
queuelib==1.6.2
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
rfc3986==1.5.0
rich==13.0.1
Scrapy==2.7.1
selenium==4.7.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
sqlparse==0.4.3
stack-data==0.6.0
tabula-py==2.6.0
tenacity==8.1.0
tldextract==3.4.0
tls-client==0.1.8
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.12
virtualenv==20.17.1
w3lib==2.1.1
wcwidth==0.2.5
Werkzeug==2.2.2
wordcloud==1.8.2.post5+gdbf7ab7
wsproto==1.2.0
WTForms==3.0.1
zope.interface==5.5.2
